requirements.txt
playwright
pytest
pytest-asyncio
allure-pytest
